Added Practical Data Science with R, Second Edition
Practical Data Science with R, Second Edition - https://livebook.manning.com/book/practical-data-science-with-r-second-edition/welcome/v-6/ - The intent of this book is to present data science from a pragmatic, practice-oriented viewpoint. The book concentrates on the process of data science, from the planning stages of a project, through the data collection and exploration, to the modeling, and finally to deployment and the sharing of results. Authors: Nina Zumel and John Mount.