Added R Tutorial - Be a Data Science rock star with R
R Tutorial – Be a Data Science rock star with R

https://data-flair.training/blogs/r-tutorial/

A tour of the R programming language that explores its different and essential concepts. This R DataFlair Tutorial Series is designed to help beginners to get started with R and experienced to brush up their R programming skills and gain perfection in the language.